Perform (hopefully safe) static cast to signed integer before compare.

+++ include/ips4o/classifier.hpp
@@ -35,6 +35,7 @@
 
 #pragma once
 
+#include <cstdint>
 #include <type_traits>
 #include <utility>
 
@@ -126,7 +127,7 @@ class Sorter<Cfg>::Classifier {
   template <bool kEqualBuckets, class Yield, int...Args>
   void classifySwitch(iterator begin, iterator end, Yield&& yield,
 		      std::integer_sequence<int, Args...>) const {
-    IPS4OML_ASSUME_NOT(log_buckets_ <= 0 && log_buckets_ >= sizeof...(Args));
+    IPS4OML_ASSUME_NOT(log_buckets_ <= 0 && log_buckets_ >= static_cast<std::int64_t>(sizeof...(Args)));
     ((Args == log_buckets_ &&
       classifyUnrolled<kEqualBuckets, Args>(begin, end, std::forward<Yield>(yield)))
      || ...);
